yolo_GTSRB/cloudbuild/requirements.txt
kagglehub
google-cloud-storage